+++ source/EngineTests/MuscleTests.cpp
@@ -66,8 +66,7 @@ TEST_F(MuscleTests, doNothing)
 
 TEST_F(MuscleTests, moveForward)
 {
-    _parameters.features.legacyModes = true;
-    _parameters.legacyCellFunctionMuscleMovementModeActivated = true;
+    _parameters.cellFunctionMuscleMovementTowardTargetedObject = false;
     _simController->setSimulationParameters(_parameters);
 
     DataDescription data;
@@ -106,8 +105,7 @@ TEST_F(MuscleTests, moveForward)
 
 TEST_F(MuscleTests, moveBackward)
 {
-    _parameters.features.legacyModes = true;
-    _parameters.legacyCellFunctionMuscleMovementModeActivated = true;
+    _parameters.cellFunctionMuscleMovementTowardTargetedObject = false;
     _simController->setSimulationParameters(_parameters);
 
     DataDescription data;
@@ -145,8 +143,7 @@ TEST_F(MuscleTests, moveBackward)
 
 TEST_F(MuscleTests, multipleMovementDirections)
 {
-    _parameters.features.legacyModes = true;
-    _parameters.legacyCellFunctionMuscleMovementModeActivated = true;
+    _parameters.cellFunctionMuscleMovementTowardTargetedObject = false;
     _simController->setSimulationParameters(_parameters);
 
     DataDescription data;